pyproject.toml
[tool.poetry]
name = "permission-manager-drf"
version = "0.1.0"
description = "A simple way to manage object permissions in drf."
authors = ["Grigory Mishchenko <grishkokot@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "permission_manager_drf"}]
repository = "https://github.com/kindlycat/permission-manager-drf"
keywords = ["permissions", "drf", "rest framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: Django",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
permission-manager = "^0.3.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.24"
ipdb = "^0.13.13"
pytest = "^8.1.2"
ruff = "^0.4.2"
pytest-django = "^4.8.0"
pytest-mock = "^3.14.0"
tox = "^4.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 79
indent-width = 4

target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", "D101", "ANN101", "D102", "EXE001", "D107", "ANN003", "ANN001",
    "ARG002", "ANN002", "D103", "D106", "D104", "ANN201", "COM812", "ISC001",
    "ANN401", "ANN102", "TRY003", "EM102", "EM101"
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "RUF012", "D105"]

[tool.ruff.lint.isort]
section-order = [
    "future", "standard-library", "third-party", "first-party", "local-folder"
]
lines-after-imports = 2
combine-as-imports = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -s"